IOTest/IOTest -- Create unified script for fio and ddr tests
Create a new IOTest script which will handle both fio and ddr tests.

Specifications for what to run can be entered either on the command
line or included in files specified on the command line (a prefix
of '@' denotes a file to read for test specs).

Specifications for fio tests are of the form:

  DDR Test Specification = ddr:$FileSz:$BlkSz:$BorD
  Fio Test Specification = fio:$IOType:$FileSz:$BlkSz:$BorD:$Engine:$IODepth

where

  IOType  = read | randread
  FileSz  = nnM ; typical: ( 2048M 1024M 512M 256M )
  BlkSz   = nnk ; typical: ( 1024k 256k 64k 16k 4k )
  BorD    = buffered | direct
  Engine  = sync | libaio
  IODepth = nn ; typical: ( 64 32 16 4 1 )

The number of concurrent jobs to test is specified
by the variable NJ defined near the end of the script;
typical: ( 1 2 3 4 5 6 7 8 10 12 14 16 18 20 22 24 )

The results for a specified run are appended to a file
with the same name as the specification w/ '.out' appended
to the end.  (A specification may have an optional '.out'
suffix that will be ignored, allowing additional runs of
an existing spec just by naming the current output file.)